pyproject.toml
[tool.poetry]
name = "postgres-seed"
version = "0.1.0"
description = ""
authors = ["PatrikTison <patriktison@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
sqlmodel = "^0.0.8"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
httpx = "^0.24.1"
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"
pandas = "^2.0.3"
black = "^23.9.1"
isort = "^5.12.0"
pylint = "^2.17.5"
pytest-mock = "^3.11.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
